+++ src/Symfony/Component/Form/DependencyInjection/FormPass.php
@@ -18,7 +18,6 @@ use Symfony\Component\DependencyInjection\Compiler\ServiceLocatorTagPass;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Exception\InvalidArgumentException;
 use Symfony\Component\DependencyInjection\Reference;
-use Symfony\Component\Form\FormTypeExtensionInterface;
 
 /**
  * Adds all services with the tags "form.type", "form.type_extension" and
@@ -95,13 +94,9 @@ class FormPass implements CompilerPassInterface
             $typeExtensionClass = $container->getParameterBag()->resolveValue($serviceDefinition->getClass());
 
             if (isset($tag[0]['extended_type'])) {
-                if (!method_exists($typeExtensionClass, 'getExtendedTypes')) {
-                    @trigger_error(sprintf('Not implementing the static getExtendedTypes() method in %s when implementing the %s is deprecated since Symfony 4.2. The method will be added to the interface in 5.0.', $typeExtensionClass, FormTypeExtensionInterface::class), E_USER_DEPRECATED);
-                }
-
                 $typeExtensions[$tag[0]['extended_type']][] = new Reference($serviceId);
                 $typeExtensionsClasses[] = $typeExtensionClass;
-            } elseif (method_exists($typeExtensionClass, 'getExtendedTypes')) {
+            } else {
                 $extendsTypes = false;
 
                 $typeExtensionsClasses[] = $typeExtensionClass;
@@ -113,8 +108,6 @@ class FormPass implements CompilerPassInterface
                 if (!$extendsTypes) {
                     throw new InvalidArgumentException(sprintf('The getExtendedTypes() method for service "%s" does not return any extended types.', $serviceId));
                 }
-            } else {
-                throw new InvalidArgumentException(sprintf('"%s" tagged services have to implement the static getExtendedTypes() method. Class "%s" for service "%s" does not implement it.', $this->formTypeExtensionTag, $typeExtensionClass, $serviceId));
             }
         }